Fix Android on-screen joystick not working at startup and pause menu Advanced Settings
Two Android-specific bugs fixed:

1. On-screen joystick enabled but non-functional at startup

   The on-screen joystick overlay appeared but touch input was ignored.
   Users had to toggle the checkbox off/on in the ImGui Virtual Keyboard
   panel to make it work.

   Root cause: At startup, inputdevice_updateconfig() ran before the
   joystick port assignment was fully resolved, leaving the input
   subsystem with stale empty mappings. The toggle path worked because
   it called import_joysticks() which forced a fresh device enumeration.

   Fix: Call import_joysticks() in the gfx_window.cpp startup path
   before enabling the on-screen joystick, matching the toggle path.
   Also removed the early-out guard in on_screen_joystick_set_enabled()
   that skipped inputdevice_config_change() when the port ID already
   matched, and added joystick_refresh_needed for the Input panel.

2. Pause menu 'Advanced Settings' button did nothing

   Pressing 'Advanced Settings' in the Android pause menu (shown on
   Back button press) was supposed to open the ImGui GUI but had no
   effect.

   Root cause: The AlertDialog caused the SDL window to lose focus.
   The injected KEYCODE_BACK event was dropped by handle_key_event()'s
   focus check (isfocus() == 0).

   Fix: Move the SDL_SCANCODE_AC_BACK handler before the focus guard,
   inside #ifdef __ANDROID__. The back button now always triggers
   AKS_ENTERGUI regardless of window focus state.
